add ShowPartitionsCommand
